+++ exinst-aeson/default.nix
@@ -1,4 +1,4 @@
-{ mkDerivation, aeson, base, constraints, exinst, singletons_2_1
+{ mkDerivation, aeson, base, constraints, exinst, singletons
 , stdenv
 }:
 mkDerivation {
@@ -6,7 +6,7 @@ mkDerivation {
   version = "0.1.0.2";
   src = ./.;
   libraryHaskellDepends = [
-    aeson base constraints exinst singletons_2_1
+    aeson base constraints exinst singletons
   ];
   homepage = "https://github.com/k0001/exinst";
   description = "Derive instances for the `aeson` library for your existential types";