+++ exinst-hashable/default.nix
@@ -1,4 +1,4 @@
-{ mkDerivation, base, constraints, exinst, hashable, singletons_2_1
+{ mkDerivation, base, constraints, exinst, hashable, singletons
 , stdenv
 }:
 mkDerivation {
@@ -6,7 +6,7 @@ mkDerivation {
   version = "0.1.0.2";
   src = ./.;
   libraryHaskellDepends = [
-    base constraints exinst hashable singletons_2_1
+    base constraints exinst hashable singletons
   ];
   homepage = "https://github.com/k0001/exinst";
   description = "Derive instances for the `hashable` library for your existential types";