+++ exinst-deepseq/CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 0.2
+
+* Depend on `singletons-2.2`, which means `KProxy` is gone.
+
+
 # Version 0.1.0.1
 
 * Relax upper bound on `base`, `singletons`, `constraints`.